Fix handling of default values for placeholders.
In order to make default placeholders work, we had to modify the fit and path outline
routines to ignore them.  That also necessitated switching back to the original
outline/ancestors logic from Drupal 7, which with a very slight tweak to the masks
and '/'-prefix on paths still works just as it should.